requirements.txt
# Requisitos para el Sistema de Puntuación de Idiomas - Tutorium
# API Backend y Herramientas de Análisis

# Framework web
Flask==2.3.3
Flask-CORS==4.0.0

# Procesamiento de texto y análisis lingüístico
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# Análisis de pronunciación y audio
librosa==0.10.1
soundfile==1.0.2
scipy==1.11.4

# Base de datos y persistencia
SQLAlchemy==2.0.23
sqlite3

# Utilidades y procesamiento de datos
numpy==1.24.4
pandas==2.1.4
requests==2.31.0

# Testing y desarrollo
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Variables de entorno
python-dotenv==1.0.0

# Logging y monitoreo
loguru==0.7.2

# Validación de datos
pydantic==2.5.0

# Análisis de sentiment y contexto
vaderSentiment==3.3.2

# Métricas y estadísticas
scikit-learn==1.3.2

# Configuración y archivos
PyYAML==6.0.1
configparser==6.0.0

# Utilidades de fecha y tiempo
python-dateutil==2.8.2

# Encriptación y seguridad
cryptography==41.0.8

# Manejo de archivos multimedia
Pillow==10.1.0

# Análisis de patrones de texto
regex==2023.10.3

# Detección de idioma
langdetect==1.0.9

# Corrección ortográfica
pyspellchecker==0.7.2

# Análisis sintáctico avanzado
stanza==1.7.0

# Formato de fechas
iso8601==2.1.0
